feat: Refactor start content and control css
