+++ src/core/components/AssistantMessage.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useState } from 'react';
 import { useIsMobile } from '../../shared/theme/useIsMobile';
 import { ClipboardCheck, Copy } from 'lucide-react';
-import styles from './ChatView.module.css';
+import styles from './Conversation.module.css';
 import KavaIcon from '../assets/KavaIcon';
 import { Content } from './Content';
 import { ThinkingContent } from './ThinkingContent';
